python/django_rest/requirements.txt
Django~=4.2
djangorestframework~=3.14.0
psycopg~=3.1.8
gunicorn~=20.1.0
python-dotenv~=1.0.0